remove bash-specific pushd/popd usage (#3)
The script's shebang is `#!/bin/sh`, so bash-specific commands such as `pushd` and `popd` are not allowed. They may work under some bash configurations, but they do not work on the default Ubuntu bash. Also, they would not work on a non-bash shell (as many Phabricator machines are likely to be running, like alpine or busybox).